Move support-bundle-reader-lib into support-bundle-viewer, outside omicron
