--- dev-tools/support-bundle-reader-lib/src/lib.rs
@@ -1,11 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-//! Utilities to inspect support bundles
-
-mod bundle_accessor;
-mod dashboard;
-mod index;
-
-pub use dashboard::run_dashboard;
